+++ AppVersion.cs
@@ -2,12 +2,8 @@ namespace MissileCamera
 {
     internal static class AppVersion
     {
-        public const string ReleaseBase = "2.32.24";
+        public const string ReleaseBase = "2.2.0";
         public const string BepInSemVer = ReleaseBase;
-        public const string VersionChannel = "DEV";
-        public const int CycleBuildNumber = 1;
-        public const string ChangeLetters = "QV";
-        public const int SubNumber = 1;
-        public const string DisplayVersion = "2.32.24QV";
+        public const string DisplayVersion = ReleaseBase;
     }
 }